+++ src/api/http.go
@@ -65,8 +65,28 @@ type muxConfig struct {
 
 // HTTPResponse represents the http response struct
 type HTTPResponse struct {
-	Error string `json:"error"`
-	Data  string `json:"data"`
+	Error *HTTPError  `json:"error,omitempty"`
+	Data  interface{} `json:"data,omitempty"`
+}
+
+// HTTPError is included in an HTTPResponse
+type HTTPError struct {
+	Message string `json:"message"`
+	Code    int    `json:"code"`
+}
+
+// NewHTTPErrorResponse returns an HTTPResponse with the Error field populated
+func NewHTTPErrorResponse(code int, msg string) HTTPResponse {
+	if msg == "" {
+		msg = http.StatusText(code)
+	}
+
+	return HTTPResponse{
+		Error: &HTTPError{
+			Code:    code,
+			Message: msg,
+		},
+	}
 }
 
 func create(host string, c Config, gateway Gatewayer) (*Server, error) {
@@ -271,8 +291,6 @@ func newServerMux(c muxConfig, gateway Gatewayer, csrfStore *CSRFStore, rpc *web
 	// get balance of addresses
 	webHandlerV1("/balance", getBalanceHandler(gateway))
 
-	webHandlerV1("/transaction/verify", verifyTxnHandler(gateway))
-
 	// Wallet interface
 
 	// Returns wallet info
@@ -384,6 +402,9 @@ func newServerMux(c muxConfig, gateway Gatewayer, csrfStore *CSRFStore, rpc *web
 	// get txn by txid
 	webHandlerV1("/transaction", getTransactionByID(gateway))
 
+	// parse and verify transaction
+	webHandlerV1("/transaction/verify", verifyTxnHandler(gateway))
+
 	// Health check handler
 	webHandlerV1("/health", healthCheck(gateway))
 
@@ -405,6 +426,8 @@ func newServerMux(c muxConfig, gateway Gatewayer, csrfStore *CSRFStore, rpc *web
 	webHandlerV1("/uxout", getUxOutByID(gateway))
 	// get all the address affected uxouts.
 	webHandlerV1("/address_uxouts", getAddrUxOuts(gateway))
+
+	webHandlerV1("/address/verify", http.HandlerFunc(addressVerify))
 
 	// Explorer handler